Changed all double variables to float to reduce memory usage in hopes of increasing speed

+++ src/sharp-eye/include/sharp-eye/point_sim.hpp
@@ -10,14 +10,14 @@ class PointSim{
 
     public:
     int n; //< Number of points
-    std::vector<Eigen::Vector3d> point_list;
-    double d; //< Size of the cube in which the points will be simulated.
+    std::vector<Eigen::Vector3f> point_list;
+    float d; //< Size of the cube in which the points will be simulated.
 
     // Transforms
-    Eigen::Transform<double,3,2> T_world2cam1;
-    Eigen::Transform<double,3,2> T_world2cam2;
-    Eigen::Transform<double,3,2> T_cam12cam2;
-    Eigen::Transform<double,3,2> T_caml2camr;
+    Eigen::Transform<float,3,2> T_world2cam1;
+    Eigen::Transform<float,3,2> T_world2cam2;
+    Eigen::Transform<float,3,2> T_cam12cam2;
+    Eigen::Transform<float,3,2> T_caml2camr;
 
     // Frames
     VisualSlamBase::Frame current_frame;
@@ -45,13 +45,13 @@ class PointSim{
      * @param camera_initial 
      * @param camera_final 
      */
-    void SetCameraPositions(Eigen::Transform<double,3,2> camera_initial,Eigen::Transform<double,3,2> camera_final);
+    void SetCameraPositions(Eigen::Transform<float,3,2> camera_initial,Eigen::Transform<float,3,2> camera_final);
 
     /**
      * @brief Sets the T_caml2camr transform
      * 
      */
-    void SetInterCameraTransform(Eigen::Transform<double,3,2> transform);
+    void SetInterCameraTransform(Eigen::Transform<float,3,2> transform);
 
     /**
      * @brief Set the initial position of the camera and the change in its position ie T_world2cam1 and T_cam22cam1
@@ -59,7 +59,7 @@ class PointSim{
      * @param camera_initial 
      * @param transform 
      */
-    void SetCameraPositionAndTransform(Eigen::Transform<double,3,2> camera_initial,Eigen::Transform<double,3,2> transform);
+    void SetCameraPositionAndTransform(Eigen::Transform<float,3,2> camera_initial,Eigen::Transform<float,3,2> transform);
     /**
      * @brief Creates a list of random 3D points enclosed by a cube of the specified size. 
      * The cube will be centered at the origin by default
@@ -67,7 +67,7 @@ class PointSim{
      * @param no_of_points 
      * @param cube_dimension 
      */
-    void CreateRandomPoints(int no_of_points,double cube_dimension);
+    void CreateRandomPoints(int no_of_points,float cube_dimension);
 
     /**
      * @brief Creates the curent and previous frames by projecting points, creating 
@@ -76,7 +76,7 @@ class PointSim{
      */
     void CreateFrames(); 
 
-    cv::KeyPoint ProjectPoints(Eigen::Vector3d point_3d,VisualSlamBase::Camera camera);
+    cv::KeyPoint ProjectPoints(Eigen::Vector3f point_3d,VisualSlamBase::Camera camera);
 
     bool InFieldOfView(cv::KeyPoint keypoint_l,cv::KeyPoint keypoint_r);
 };
